move progress-extract into cli folder
