+++ cli/progress-extract/bar.js
