+++ cli/progress-extract/index.js
