+++ package.json
@@ -20,6 +20,7 @@
     "commander": "^9.4.0",
     "conf": "^10.2.0",
     "enquirer": "^2.3.6",
+    "extract-zip": "^2.0.1",
     "filesize": "^10.0.7",
     "form-data": "^4.0.0",
     "glob": "^8.0.3",
